GHouses remove logic fix

+++ src/server/scripts/Custom/Guild/mod_guild_points.cpp
@@ -329,9 +329,12 @@ bool sModGuildPoints::AddGuildHousePoints(Player* player, uint32 points)
     return true;
 }
 
-void sModGuildPoints::DeleteGuild(uint32 guildId)
+void sModGuildPoints::DeleteGuild(uint32 guildId, bool removeRanking)
 {
-    CharacterDatabase.PExecute("DELETE FROM guild_points_ranking WHERE guildId = %u;", guildId);
+    if (removeRanking)
+    {
+        CharacterDatabase.PExecute("DELETE FROM guild_points_ranking WHERE guildId = %u;", guildId);
+    }
     // Delete actual guild_house data from characters database
     CharacterDatabase.PExecute("DELETE FROM `guild_house` WHERE `guild` = '%u'", guildId);
     CharacterDatabase.PExecute("DELETE FROM `guild_house_purchased_spawns` WHERE `guild` = '%u'", guildId);
